Move adapters specific for Arduino to a dedicated directory.
