+++ lib/3rd_party_adapters/Arduino/SerialOutputStreamBuffer.hpp
