+++ lib/3rd_party_adapters/Arduino/framework_adapter.cpp
